+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Collected winning data on auction tab
 - Proposal cell display
 - Tooltip display
+- Error message on Bridge tokens form
 
 #### Removed
 - Scrolling when the drover is open